feat: fold fpc into acc_store wrappers

+++ test/vpto/cases/micro-op/cv-mixed/fixpipe-quant-ub-cv/launch.cpp
@@ -39,12 +39,12 @@ struct MrgSortExecutedNumList {
 #endif
 
 extern "C" __global__ [aicore] void fixpipe_quant_ub_cv_kernel(
-    __gm__ __fp16 *a, __gm__ __fp16 *b, __gm__ float *fp, __gm__ __fp16 *out);
+    __gm__ __fp16 *a, __gm__ __fp16 *b, __gm__ uint32_t *fp, __gm__ __fp16 *out);
 
-void LaunchFixpipe_quant_ub_cv_kernel(__fp16 *a, __fp16 *b, float *fp,
+void LaunchFixpipe_quant_ub_cv_kernel(__fp16 *a, __fp16 *b, uint32_t *fp,
                                       __fp16 *out, void *stream) {
   fixpipe_quant_ub_cv_kernel<<<1, nullptr, stream>>>((__gm__ __fp16 *)a,
                                                      (__gm__ __fp16 *)b,
-                                                     (__gm__ float *)fp,
+                                                     (__gm__ uint32_t *)fp,
                                                      (__gm__ __fp16 *)out);
 }
